\name{get3primeIVTprobefileData}
\alias{get3primeIVTprobefileData}
\title{Read a data file describing the probe sequences on an Affymetrix genechip}
\description{Read a data file describing the probe sequences on an Affymetrix genechip}
\usage{
get3primeIVTprobefileData(arraytype, datafile, pkgname, chip.pd, comparewithcdf = FALSE)
}
\arguments{
  \item{arraytype}{Character. Array type (e.g. 'HG-U133A')}
  \item{datafile}{Character. The filename of the input data file,
    or a connection (see example).
    If omitted a default name is constructed from \code{arraytype}
    (for details you will need to consult this function's source code).}
  \item{pkgname}{Character. Package name. If NULL the name is derived
    from \code{arraytype}.}
  \item{chip.pd}{Character. Name of the platform design file for the \code{arraytype}.}
  \item{comparewithcdf}{Logical. If TRUE, run a consistency check
    against a CDF package of the same name (what used to be Laurent's
    "extraparanoia".)}
}
\details{This function serves as an interface between
  the (1) representation of array probe information data in the packages
  that are generated by \code{makeProbePackageGCSs} and (2) the
  vendor- and possibly version-specific way the data are represented
  in \code{datafile}.

  \code{datafile} is a tabulator-separated file with one row per probe,
  and column names \code{'probesetid'}, \code{'fsetid'},
  \code{'fid'}, \code{'x'}, \code{'y'}, and \code{'GC.count'}. See the vignette
  for an example.
}

\value{A list with three components
    \item{dataEnv}{an environment which contains the data frame with the
      probe sequences and the other probe data.}
    \item{symVal}{a named list of symbol value substitutions which can
      be used to customize the man pages. See
      \code{\link[Biobase:createPackage]{createPackage}}.}
    \item{pkgname}{a character with the package name; will be the same
      as the function parameter \code{pkgname} if it was specified;
      otherwise, the name is constructed from the parameter \code{arraytype}.}
}

\seealso{makeProbePackageGCSs}
\examples{
if (length(list.files(path = ".", pattern = "*.CEL")) != 0){
  ## Example using the "Mouse430 2.0" chip-type
  
  ## Input the clean name for the given chip:
  chip <- "mouse4302"

  ## Input the .probe_tab file, as generated using the internal function:
     ##  IVT3primepFBuilder()
  
  probedata <- "GCSs.mouse4302.probeFile.probe_tab"
  
  ## Run function:
  get3primeIVTprobefileData(arraytype = chip, datafile = probedata)
  }
}
\keyword{IO}
\keyword{utilities}
